reverted wren changes, cleaned up env vars, fixed sparrow

+++ tests/integration/api/query_v1_panic_test.go
@@ -56,6 +56,9 @@ var _ = Describe("Query V1 when Sparrow panics", func() {
 	})
 
 	AfterEach(func() {
+		if os.Getenv("LOCAL") == "true" {
+			Skip("tests running locally, skipping sparrow panic test")
+		}
 		// delete table
 		_, err := tableClient.DeleteTable(ctx, &v1alpha.DeleteTableRequest{TableName: tableName})
 		Expect(err).ShouldNot(HaveOccurredGrpc())
